+++ src/components/PokemonDetails.js
